pyproject.toml
[tool.poetry]
name = "sdss-lvmieb"
version = "0.4.0-alpha.0"
description = "Control software for the Local Volume Mapper Instrument Electronics Box"
authors = ["Changgon Kim <changgonkim@khu.ac.kr>"]
maintainers = ["José Sánchez-Gallego <gallegoj@uw.edu>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://github.com/sdss/lvmieb"
repository = "https://github.com/sdss/lvmieb"
documentation = "https://sdss-lvmieb.readthedocs.org"
keywords = ["astronomy", "software"]
classifiers = [
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
	"Programming Language :: Python :: 3.9",
    "Topic :: Documentation :: Sphinx",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
    { include = "lvmieb", from = "python" }
]
include = ["python/lvmieb/etc/*"]

[tool.poetry.scripts]
lvmieb = "lvmieb.__main__:lvmieb"

[tool.poetry.dependencies]
python = "^3.9,<3.11"
sdsstools = ">=0.4.0"
sdss-clu = "^1.6.1"
click-default-group = "^1.2.2"
numpy = "^1.20.3"
sdss-drift = "^0.4.2"

[tool.poetry.dev-dependencies]
ipython = ">=7.11.0"
flake8 = ">=3.7.9"
doc8 = ">=0.8.0"
pytest = ">=5.2.2"
pytest-asyncio = ">=0.10.0"
pytest-cov = ">=2.8.1"
pytest-mock = ">=1.13.0"
pytest-sugar = ">=0.9.2"
isort = ">=4.3.21"
codecov = ">=2.0.15"
coverage = {version = ">=5.0", extras = ["toml"]}
ipdb = ">=0.12.3"
rstcheck = "^3.3.1"
Sphinx = ">=3.0.0"
black = ">=21.7b0"
sphinx-click = ">=2.6.0"
sphinx-jsonschema = ">=1.16.7"
myst-parser = ">=0.14.0"
furo = "^2021.6.18-beta.36"
nox = "^2021.6.12"
sphinx-autobuild = "^2021.3.14"
sphinx-copybutton = "^0.3.3"

[tool.isort]
line_length = 79
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "SDSS", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"
known_first_party = "lvmieb"
known_sdss = ["sdsstools", "clu", "drift"]
balanced_wrapping = true
include_trailing_comma = false
lines_after_imports = 2
use_parentheses = true

[tool.pytest.ini_options]
addopts = "--cov lvmieb --cov-report xml --cov-report html --cov-report term"
asyncio_mode = "auto"

[tool.black]
line-length = 88
target-version = ['py39']
fast = true

[tool.coverage.run]
branch = true
include = ["python/lvmieb/*"]
omit = [
    "python/lvmieb/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "# pragma: no cover",
    "if TYPE_CHECKING"
]

[build-system]
requires = ["poetry-core>=1.1.0a7"]
build-backend = "poetry.core.masonry.api"
